+++ src/rcpp_get_bearing.cpp
@@ -19,7 +19,7 @@ using namespace Rcpp;
 //' @aliases rcpp_get_bearing
 //' @rdname rcpp_get_bearing
 //'
-//' @export
+//' @keywords internal
 // [[Rcpp::export]]
 double rcpp_get_bearing(double x1, double y1, double x2, double y2) {